Squashed 'vision' changes from 725c5d6..d20248928 (#1032)
d20248928 Merge pull request #688 from gouthamb/next-aiedev2

+++ vision/L1/examples/blacklevel/xf_black_level_accel.cpp
@@ -30,8 +30,8 @@ static constexpr int __XF_DEPTH =
 
 void blackLevelCorrection_accel(ap_uint<IMAGE_PTR_WIDTH>* in_img_ptr,
                                 ap_uint<IMAGE_PTR_WIDTH>* out_img_ptr,
-                                int black_level,
-                                float mul_value,
+                                uint32_t black_level,
+                                int mul_value,
                                 int height,
                                 int width) {
 // -----------------------------------------------
